jasper: use the GLUT framework on macOS
This was breaking Qt. I’m not sure using `darwin` as a parameter here is
right; maybe this would have to move back out of `pkgs/by-name`? 

There are probably more cases like this; it might be good to have a
top‐level glut package that resolves to either GLUT or freeglut
depending on platform or something?